Try to use RefCell to make lambda calls smaller.
I want to call my lambdas without specifying any parameters. So the
things, that the lambda wants to change have to be capured by (mutable)
reference. But then these things are not only caputered, when the lambda
is called, but as long as the lambda exists. So as long as the lambda
exists, I cannot change the things outside of the lambda, which is very
inconvinient. One solution is a RefCell.

But using RefCells is very ugly, I see now.

I will try to make it more nice.