Fix: Let "cargo run" work without option "--bin"
Because we have multiple binaries, that "cargo run" can run, we must
specify the binary with "--bin <binary>", unless we specify a default
binary in the manifest.

The default binary is now the binary, whose source code is in
src/main.rs. Now we can run "cargo run" as described in in readme file.